Add styling to GUI using CSS and refactor DialogBox classes
CSS styles adopted from https://github.com/rebeccacxy/ip. DialogBox classes refactored into 2 separate dialog box classes, as the flip function was unnecessary since we can simply abstract the UI for the dialog boxes for chatbot and user into separate files.
