+++ src/main/resources/view/ChatbotDialogBox.fxml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.HBox?>
+<fx:root xmlns:fx="http://javafx.com/fxml/1" alignment="CENTER_RIGHT" maxHeight="Infinity" maxWidth="Infinity"
+         prefWidth="400.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171">
+    <ImageView fx:id="displayPicture" fitHeight="50.0" fitWidth="50.0" pickOnBounds="true" preserveRatio="true"/>
+    <HBox fx:id="speechBubble" alignment="CENTER_RIGHT" styleClass="speech-bubble">
+        <HBox.margin>
+            <Insets bottom="5.0" left="10.0" right="10.0" top="5.0"/>
+        </HBox.margin>
+        <Label fx:id="dialog" text="Label" wrapText="true" HBox.hgrow="ALWAYS">
+            <padding>
+                <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
+            </padding>
+            <HBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" top="5.0"/>
+            </HBox.margin>
+        </Label>
+    </HBox>
+    <padding>
+        <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+    </padding>
+</fx:root>
