+++ src/main/resources/view/MainWindow.fxml
@@ -8,17 +8,13 @@
 <AnchorPane xmlns:fx="http://javafx.com/fxml/1" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity"
             minWidth="-Infinity"
             prefHeight="600.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/8.0.171"
-            fx:controller="dukeegg.MainWindow">
-    <children>
-        <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="324.0"
-                   AnchorPane.bottomAnchor="1.0"/>
-        <Button fx:id="sendButton" layoutX="324.0" layoutY="558.0" mnemonicParsing="false" onAction="#handleUserInput"
-                prefHeight="41.0" prefWidth="76.0" text="Send"/>
-        <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="557.0" prefWidth="400.0"
-                    vvalue="1.0">
-            <content>
-                <VBox fx:id="dialogContainer" prefHeight="552.0" prefWidth="388.0"/>
-            </content>
-        </ScrollPane>
-    </children>
+            fx:controller="ui.MainWindow">
+    <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="324.0"
+               AnchorPane.bottomAnchor="1.0" promptText="Enter a command, type help to find out more!"/>
+    <Button fx:id="sendButton" layoutX="324.0" layoutY="558.0" mnemonicParsing="false" onAction="#handleUserInput"
+            prefHeight="41.0" prefWidth="76.0" text="Send"/>
+    <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="557.0" prefWidth="400.0"
+                vvalue="1.0">
+        <VBox fx:id="dialogContainer" prefHeight="552.0" prefWidth="388.0"/>
+    </ScrollPane>
 </AnchorPane>
